fix binary/logical and
in update & delete functions the if statement was using binary and which seems incorrect. Fixing this to logical and.

+++ src/lib/mib_cache.php
@@ -516,7 +516,7 @@ class MibCache{
 			/* get list of columns for this mib table */
 			$columns = $this->cache__tables_columns[$this->active_mib][$this->active_table];
 
-			if ($columns & cacti_sizeof($columns) > 0) {
+			if ($columns && cacti_sizeof($columns) > 0) {
 				foreach ($columns as $column_params) {
 					$column_params['oid'] .= '.' . $this->active_table_entry;
 					db_execute_prepared('DELETE FROM `snmpagent_cache` WHERE `oid` = ?', array($column_params['oid']));
@@ -545,7 +545,7 @@ class MibCache{
 		if ($oid_entry !== false) {
 			$columns = $this->cache__tables_columns[$this->active_mib][$this->active_table];
 
-			if ($columns & cacti_sizeof($columns) > 0) {
+			if ($columns && cacti_sizeof($columns) > 0) {
 				$sql = array();
 
 				foreach ($columns as $column_params) {
